+++ ErsatzTV.Infrastructure/Search/Models/ElasticSearchItem.cs
@@ -117,6 +117,9 @@ public class ElasticSearchItem : MinimalElasticSearchItem
 
     [JsonPropertyName(LuceneSearchIndex.ShowStudioField)]
     public List<string> ShowStudio { get; set; }
+    
+    [JsonPropertyName(LuceneSearchIndex.ShowContentRatingField)]
+    public List<string> ShowContentRating { get; set; }
 
     [JsonPropertyName(LuceneSearchIndex.StyleField)]
     public List<string> Style { get; set; }